Firewall-Log-Analyzer/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
dnspython==1.16.0
fastapi==0.127.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==3.12.0
python-dotenv==1.2.1
requests==2.31.0
reportlab==4.4.4
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
sendgrid==6.11.0
redis==5.0.1
# ML Engine Dependencies
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
imbalanced-learn==0.12.3
